# tests/openapi_gen.json
{
  "openapi": "3.0.2",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
    "/api/users": {
      "post": {
        "summary": "Create User",
        "operationId": "create_user_api_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserCreate" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/token": {
      "post": {
        "summary": "Generate Token",
        "operationId": "generate_token_api_token_post",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Type" },
            "name": "type",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_generate_token_api_token_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "summary": "Get User",
        "operationId": "get_user_api_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/User" } }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/tasks": {
      "get": {
        "summary": "Get Tasks",
        "operationId": "get_tasks_api_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Tasks Api Tasks Get",
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      },
      "post": {
        "summary": "Create Task",
        "operationId": "create_task_api_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TaskCreate" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Task" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "summary": "Get Task",
        "operationId": "get_task_api_tasks__task_id__get",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Task Id", "type": "integer" },
            "name": "task_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      },
      "put": {
        "summary": "Update Task",
        "operationId": "update_task_api_tasks__task_id__put",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Task Id", "type": "integer" },
            "name": "task_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TaskCreate" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      },
      "delete": {
        "summary": "Delete Task",
        "operationId": "delete_task_api_tasks__task_id__delete",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "Task Id", "type": "integer" },
            "name": "task_id",
            "in": "path"
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/users/{user_id}": {
      "put": {
        "summary": "Update Password",
        "operationId": "update_password_api_users__user_id__put",
        "parameters": [
          {
            "required": true,
            "schema": { "title": "User Id", "type": "integer" },
            "name": "user_id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChangePassword" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/companies": {
      "post": {
        "summary": "Create Company",
        "operationId": "create_company_api_companies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CompanyCreate" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/companies/me": {
      "get": {
        "summary": "Get Company",
        "operationId": "get_company_api_companies_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Company" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/companies/names": {
      "get": {
        "summary": "Get All Companies",
        "operationId": "get_all_companies_api_companies_names_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        }
      }
    },
    "/api/employees": {
      "post": {
        "summary": "Create Employee",
        "operationId": "create_employee_api_employees_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/_EmployeeBase" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/employees/tasks": {
      "get": {
        "summary": "Get Employees Tasks",
        "operationId": "get_employees_tasks_api_employees_tasks_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/ask_gpt": {
      "post": {
        "summary": "Ask Gpt",
        "operationId": "ask_gpt_ask_gpt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatGPTRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChatGPTResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_generate_token_api_token_post": {
        "title": "Body_generate_token_api_token_post",
        "required": ["username", "password"],
        "type": "object",
        "properties": {
          "grant_type": {
            "title": "Grant Type",
            "pattern": "password",
            "type": "string"
          },
          "username": { "title": "Username", "type": "string" },
          "password": { "title": "Password", "type": "string" },
          "scope": { "title": "Scope", "type": "string", "default": "" },
          "client_id": { "title": "Client Id", "type": "string" },
          "client_secret": { "title": "Client Secret", "type": "string" }
        }
      },
      "ChangePassword": {
        "title": "ChangePassword",
        "required": ["email", "password", "new_password", "confirm_password"],
        "type": "object",
        "properties": {
          "email": { "title": "Email", "type": "string" },
          "password": { "title": "Password", "type": "string" },
          "new_password": { "title": "New Password", "type": "string" },
          "confirm_password": { "title": "Confirm Password", "type": "string" }
        }
      },
      "ChatGPTRequest": {
        "title": "ChatGPTRequest",
        "required": ["question"],
        "type": "object",
        "properties": { "question": { "title": "Question", "type": "string" } }
      },
      "ChatGPTResponse": {
        "title": "ChatGPTResponse",
        "required": ["response"],
        "type": "object",
        "properties": { "response": { "title": "Response", "type": "string" } }
      },
      "Company": {
        "title": "Company",
        "required": ["email", "id"],
        "type": "object",
        "properties": {
          "email": { "title": "Email", "type": "string" },
          "id": { "title": "Id", "type": "integer" }
        }
      },
      "CompanyCreate": {
        "title": "CompanyCreate",
        "required": ["email", "hashed_password", "name"],
        "type": "object",
        "properties": {
          "email": { "title": "Email", "type": "string" },
          "hashed_password": { "title": "Hashed Password", "type": "string" },
          "name": { "title": "Name", "type": "string" }
        }
      },
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationError" }
          }
        }
      },
      "Task": {
        "title": "Task",
        "required": ["state", "title", "description", "priority", "id", "owner_id"],
        "type": "object",
        "properties": {
          "state": { "title": "State", "type": "string" },
          "title": { "title": "Title", "type": "string" },
          "description": { "title": "Description", "type": "string" },
          "priority": { "title": "Priority", "type": "string" },
          "id": { "title": "Id", "type": "integer" },
          "owner_id": { "title": "Owner Id", "type": "integer" }
        }
      },
      "TaskCreate": {
        "title": "TaskCreate",
        "required": ["state", "title", "description", "priority"],
        "type": "object",
        "properties": {
          "state": { "title": "State", "type": "string" },
          "title": { "title": "Title", "type": "string" },
          "description": { "title": "Description", "type": "string" },
          "priority": { "title": "Priority", "type": "string" }
        }
      },
      "User": {
        "title": "User",
        "required": ["email", "id"],
        "type": "object",
        "properties": {
          "email": { "title": "Email", "type": "string" },
          "id": { "title": "Id", "type": "integer" }
        }
      },
      "UserCreate": {
        "title": "UserCreate",
        "required": ["email", "hashed_password", "name"],
        "type": "object",
        "properties": {
          "email": { "title": "Email", "type": "string" },
          "hashed_password": { "title": "Hashed Password", "type": "string" },
          "name": { "title": "Name", "type": "string" }
        }
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": ["loc", "msg", "type"],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] }
          },
          "msg": { "title": "Message", "type": "string" },
          "type": { "title": "Error Type", "type": "string" }
        }
      },
      "_EmployeeBase": {
        "title": "_EmployeeBase",
        "required": ["company_id", "user_email"],
        "type": "object",
        "properties": {
          "company_id": { "title": "Company Id", "type": "integer" },
          "user_email": { "title": "User Email", "type": "string" }
        }
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "/api/token" } }
      }
    }
  }
}